+++ lib/Template.js
@@ -402,11 +402,8 @@ class Template {
 			"/******/ ",
 			new ConcatSource(
 				"function(__webpack_require__) { // webpackRuntimeModules\n",
-				'\t"use strict";\n\n',
-				new PrefixSource(
-					"\t",
-					this.renderRuntimeModules(runtimeModules, renderContext)
-				),
+				'"use strict";\n\n',
+				this.renderRuntimeModules(runtimeModules, renderContext),
 				"}\n"
 			)
 		);
