+++ lib/javascript/ArrayPushCallbackChunkFormatPlugin.js
@@ -9,8 +9,8 @@ const { ConcatSource, PrefixSource, RawSource } = require("webpack-sources");
 const { RuntimeGlobals } = require("..");
 const HotUpdateChunk = require("../HotUpdateChunk");
 const Template = require("../Template");
-const { generateEntryStartup } = require("../web/JsonpHelpers");
 const { getCompilationHooks } = require("./JavascriptModulesPlugin");
+const { generateEntryStartup } = require("./StartupHelpers");
 
 /** @typedef {import("../Compiler")} Compiler */
 
@@ -89,13 +89,7 @@ class ArrayPushCallbackChunkFormatPlugin {
 								);
 								if (runtimeModules.length > 0) {
 									runtime.add(
-										new PrefixSource(
-											"\t",
-											Template.renderRuntimeModules(
-												runtimeModules,
-												renderContext
-											)
-										)
+										Template.renderRuntimeModules(runtimeModules, renderContext)
 									);
 								}
 								if (entries.length > 0) {
