+++ lib/javascript/CommonJsChunkFormatPlugin.js
@@ -8,11 +8,11 @@
 const { ConcatSource, RawSource } = require("webpack-sources");
 const RuntimeGlobals = require("../RuntimeGlobals");
 const Template = require("../Template");
-const { generateEntryStartup } = require("../web/JsonpHelpers");
 const {
 	getChunkFilenameTemplate,
 	getCompilationHooks
 } = require("./JavascriptModulesPlugin");
+const { generateEntryStartup } = require("./StartupHelpers");
 
 /** @typedef {import("../Compiler")} Compiler */
