+++ lib/javascript/StartupHelpers.js
