+++ lib/runtime/OnChunksLoadedRuntimeModule.js
@@ -8,7 +8,7 @@ const RuntimeGlobals = require("../RuntimeGlobals");
 const RuntimeModule = require("../RuntimeModule");
 const Template = require("../Template");
 
-class ChunkLoadedRuntimeModule extends RuntimeModule {
+class OnChunksLoadedRuntimeModule extends RuntimeModule {
 	constructor() {
 		super("chunk loaded");
 	}
@@ -68,4 +68,4 @@ class ChunkLoadedRuntimeModule extends RuntimeModule {
 	}
 }
 
-module.exports = ChunkLoadedRuntimeModule;
+module.exports = OnChunksLoadedRuntimeModule;
